requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
deepgram-sdk==3.9.0
httpx==0.28.1
openai==1.61.1
assemblyai==0.37.0
openpyxl==3.1.5
fpdf==1.7.2
python-docx==1.1.2
nltk==3.9.1
transformers==4.48.1
fsspec==2024.9.0  # transformers dependency pin for deterministic installs
google-cloud-translate==3.19.0
google-auth==2.38.0
requests==2.32.3
gunicorn==23.0.0
pytest==8.3.4
